Don't immediately request completions in eglot-completion-at-point
Yet another adjustment to this function.  According to the
documentation of completion-at-point-functions, we should strive to
make functions like eglot-completion-at-point "cheap to run".
Requesting completion from the server immediately after calling the
function goes against that.  The reason we were doing it is that it
might have helped compute more accurate "bounds" for the return value
(START and END) from possible TextEdit completion items.  But I've
decided it's not worth the effort, at least for now.

* eglot.el (eglot-completion-at-point): Request completions
asynchronously.