Per #50: use  completionItem/resolve more abundantly
It was already used to resolve documentation bits of completions, but
it can also be useful to resolve snippet templates and such.

To resolve a completion, you need some part of a completion to start
with.  If it has a :data field exists and the server supports
:resolveProvider, fetch the new object, otherwise use whatever we had
already.

* eglot.el (eglot-completion-at-point): Add another local
function for resolving completions.

#50: https://github.com/joaotavora/eglot/issues/50