Close #505: ensure completion terminates in correct buffer
To design a completion-in-region-function replacement that leverages
the elements in completion-at-point-functions, we must ensure that
their :exit-function parts execute in the correct buffer.  That is the
buffer where the text to be completed lives, not necessarily the
buffer being used for user interaction.

Later on, this guarantee should be provided by Emacs itself, perhaps
by putting the correct with-current-buffer call in completion--done.

Copyright-paperwork-exempt: yes
Co-authored-by: João Távora <joaotavora@gmail.com>

* eglot.el (eglot-completion-at-point): Ensure :exit-function's
buffer is where the source is.

#505: https://github.com/joaotavora/eglot/issues/505