Per #667: add a passing test demonstrating clangd + TRAMP works
... It works at least within the minimal, well-controlled reproducible
settings of this test.  Maybe if we knew something more about the
setup of the user who submitted this report we would be able to
concoct a failing test, but we don't.

* eglot-tests.el (subr-x): Require it
(eglot--make-file-or-dir): Return expanded file name.
(eglot-tests--lsp-abiding-column-1): New helper.
(eglot-lsp-abiding-column): Use it.
(eglot--tramp-test): Fix `skip-unless` condition.
(eglot--tramp-test-2): New test.

#667: https://github.com/joaotavora/eglot/issues/667