Fix #633, #1067: Don't exceed max-specdl-size in big Go projects
When invoking client/registerCapability for
workspace/didChangeWatchedFiles, Gopls lists each file to watch
separately.  This makes eglot--glob-emit-{} emit a closure with an
'or' form containing a potentially large number of 're-search-forward'
forms.

For large Go project such as "Kubernetes", this list becomes so large
that -- for some reason I don't understand -- it triggers the
'max-specdl-size' limit.

An alternative using `regexp` opt doesn't seem to trigger the error.

* eglot.el (eglot--glob-emit-{}): Use regexp-opt.

#633: https://github.com/joaotavora/eglot/issues/633
#1067: https://github.com/joaotavora/eglot/issues/1067